+++ includes/PF_FormPrinter.php
@@ -1070,7 +1070,7 @@ END;
 
 					$section_name = trim( $tag_components[1] );
 					$page_section_in_form = PFPageSection::newFromFormTag( $tag_components, $user );
-					$section_text = null;
+					$section_text = '';
 
 					// Split the existing page contents into the textareas in the form.
 					$default_value = "";